requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
python-dateutil==1.5
requests==2.27.0
urllib3==1.26.7
whitenoise==5.3.0
Django==4.0
django-cors-headers==3.10.0
djangorestframework==3.13.1
dj-database-url==0.5.0
django-heroku==0.3.1
psycopg2==2.9.3
psycopg==3.0.7